+++ src/main/java/com/stool/studentcooperationtools/domain/slide/Slide.java
@@ -1,7 +1,7 @@
 package com.stool.studentcooperationtools.domain.slide;
 
 import com.stool.studentcooperationtools.domain.BaseTimeEntity;
-import com.stool.studentcooperationtools.domain.presentaion.Presentation;
+import com.stool.studentcooperationtools.domain.presentation.Presentation;
 import com.stool.studentcooperationtools.domain.script.Script;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
